Front implemented generic confirm dialog

+++ frontend/src/app/components/home/contacts/contact-card/contact-card.component.ts
@@ -1,6 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
+import { GenericConfirmDialogComponent } from 'src/app/components/generic-confirm-dialog/generic-confirm-dialog.component';
 import { Contact } from 'src/app/models/contact';
+import { GenericConfirmDialog } from 'src/app/models/helpers/generic-confirm-dialog';
 import { ContactService } from 'src/app/services/contact.service';
 import { ContactCreateEditModalComponent } from '../contact-create-edit-modal/contact-create-edit-modal.component';
 
@@ -29,7 +31,20 @@ export class ContactCardComponent implements OnInit {
   }
 
   deleteContact(): void {
-    this.contactService.deleteContact(this.contact._id)
-      .subscribe(res => console.log(res));
+    this.dialog.open<GenericConfirmDialogComponent, GenericConfirmDialog, boolean>(GenericConfirmDialogComponent, {
+      data: {
+        label: 'Delete contact',
+        description: '¿Do you really want to remove the contact?',
+        firstButtonLabel: 'Cancel',
+        secondButtonLabel: 'Yes'
+      }
+    })
+      .afterClosed()
+      .subscribe(resDialog => {
+        if (resDialog) {
+          this.contactService.deleteContact(this.contact._id)
+            .subscribe(res => console.log(res));
+        }
+      });
   }
 }
